Fixed compilation with GCC
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@9382 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/race/race_manager.cpp
@@ -220,7 +220,7 @@ void RaceManager::computeRandomKartList()
     }
 
     m_ai_kart_list.clear();
-    unsigned int m = std::min( (unsigned) n,  m_default_ai_list.size());
+    unsigned int m = std::min( (unsigned) n,  (unsigned)m_default_ai_list.size());
 
     for(unsigned int i=0; i<m; i++)
     {
